Add parameter to select subsets of modules to be loaded.
Plugin groups are selected by supplyint a bit mask parameter
to the load-all-plugins() method.

Plugins are distributed in sub-directories off a root directory.
The plugin_manager is supplied with the root directory and adds
sub-directories to the search path depending on which plugin types
are selected. This shortens the default plugin paths quite a bit.

To make this approach work, the build/install paths were changed
to group like plugins into separate directories.